Refactor: Split out EditableFileEntry
