+++ packages/neoFrontend/src/pages/VizPage/Body/Editor/BottomButtons/NewTop.js
@@ -1,0 +1,9 @@
+import React from 'react';
+import { TopList, TopListItem } from './styles';
+
+export const NewTop = () => (
+  <TopList>
+    <TopListItem>new file</TopListItem>
+    <TopListItem>new directory</TopListItem>
+  </TopList>
+);
