+++ packages/neoFrontend/src/pages/VizPage/Body/Editor/FilesSection/styles.js
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import { Clickable } from '../../../../../styles';
 
-const FileStyle = props => `
+export const FileStyle = props => `
   display: flex;
   align-items: center;
   height: ${props.theme.editorEntryHeight}px;
@@ -14,26 +14,4 @@ const FileStyle = props => `
 export const FileEntry = styled(Clickable)`
   ${FileStyle}
   text-decoration: ${props => (props.isActive ? 'underline' : 'none')};
-`;
-
-export const EditableFileEntry = styled.div`
-  position: relative;
-  height: ${props => props.theme.editorEntryHeight}px;
-`;
-
-export const EditableFileInput = styled.input`
-  ${FileStyle}
-  font-family: ${props => props.theme.defaultCodingFontFamily};
-  font-size: ${props => props.theme.defaultCodingFontSize};
-  border-top: 0;
-  border-bottom: 0;
-  border-right: 0;
-  outline: 0;
-  padding-top: 0;
-  padding-bottom: 0;
-  padding-right: 0;
-  color: ${props => props.theme.dark};
-  position: absolute;
-  width: 100%;
-  box-sizing: border-box;
 `;
